# service/docs/swagger.yaml
basePath: /api/v1
definitions:
  artistrest.ChangeArtistRequestBody:
    properties:
      name:
        maxLength: 130
        type: string
    type: object
  artistrest.ChangeArtistResponse:
    properties:
      id:
        type: integer
      name:
        maxLength: 130
        type: string
    required:
    - id
    - name
    type: object
  artistrest.CreateArtistRequest:
    properties:
      name:
        maxLength: 130
        type: string
    required:
    - name
    type: object
  artistrest.CreateArtistResponse:
    properties:
      id:
        type: integer
      name:
        maxLength: 130
        type: string
    required:
    - id
    - name
    type: object
  artistrest.GetArtistResponse:
    properties:
      id:
        type: integer
      name:
        maxLength: 130
        type: string
    required:
    - id
    - name
    type: object
  artistrest.RemoveArtistResponse:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  models.ArtistAPI:
    properties:
      id:
        type: integer
      name:
        maxLength: 130
        type: string
    required:
    - id
    - name
    type: object
  models.ArtistIDAPI:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  models.Pagination:
    properties:
      pageNumber:
        minimum: 1
        type: integer
      pageSize:
        maximum: 100
        minimum: 10
        type: integer
    type: object
  models.PaginationMetadataAPI:
    properties:
      currentPageNumber:
        type: integer
      pageCount:
        type: integer
      pageSize:
        type: integer
      recordCount:
        type: integer
    type: object
  models.SongAPI:
    properties:
      artist:
        $ref: '#/definitions/models.ArtistAPI'
      id:
        type: integer
      link:
        type: string
      name:
        maxLength: 130
        type: string
      releaseDate:
        type: string
      text:
        type: string
    required:
    - id
    - link
    - name
    - releaseDate
    - text
    type: object
  mwerror.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  songrest.ChangeSongRequestBody:
    properties:
      artist:
        $ref: '#/definitions/models.ArtistIDAPI'
      link:
        type: string
      name:
        maxLength: 130
        type: string
      releaseDate:
        type: string
      text:
        type: string
    type: object
  songrest.ChangeSongResponse:
    properties:
      artist:
        $ref: '#/definitions/models.ArtistAPI'
      id:
        type: integer
      link:
        type: string
      name:
        maxLength: 130
        type: string
      releaseDate:
        type: string
      text:
        type: string
    required:
    - id
    - link
    - name
    - releaseDate
    - text
    type: object
  songrest.CreateSongRequest:
    properties:
      artist:
        $ref: '#/definitions/models.ArtistIDAPI'
      link:
        type: string
      name:
        maxLength: 130
        type: string
      releaseDate:
        type: string
      text:
        type: string
    required:
    - link
    - name
    - releaseDate
    - text
    type: object
  songrest.CreateSongResponse:
    properties:
      artist:
        $ref: '#/definitions/models.ArtistAPI'
      id:
        type: integer
      link:
        type: string
      name:
        maxLength: 130
        type: string
      releaseDate:
        type: string
      text:
        type: string
    required:
    - id
    - link
    - name
    - releaseDate
    - text
    type: object
  songrest.GetSongResponse:
    properties:
      pagination:
        $ref: '#/definitions/models.PaginationMetadataAPI'
      song:
        $ref: '#/definitions/models.SongAPI'
    type: object
  songrest.RemoveSongResponse:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  songrest.SearchSongsResponse:
    properties:
      pagination:
        $ref: '#/definitions/models.PaginationMetadataAPI'
      songs:
        items:
          $ref: '#/definitions/models.SongAPI'
        type: array
    type: object
info:
  contact: {}
  description: Микросервис библиотеки песен.
  title: Song-library-service
paths:
  /artists/:
    post:
      consumes:
      - application/json
      description: Добавить нового исполнителя. Название исполнителя должно быть уникальным.
      parameters:
      - description: Данные нового исполнителя
        in: body
        name: artist
        required: true
        schema:
          $ref: '#/definitions/artistrest.CreateArtistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/artistrest.CreateArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Добавить нового исполнителя.
      tags:
      - artist
  /artists/{artist-id}:
    delete:
      consumes:
      - application/json
      description: Удалить данные исполнителя.
      parameters:
      - in: path
        name: artist-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/artistrest.RemoveArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Удалить данные исполнителя.
      tags:
      - artist
    get:
      consumes:
      - application/json
      description: Получить данные определенного исполнителя.
      parameters:
      - in: path
        name: artist-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/artistrest.GetArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Получить данные определенного исполнителяяяя.
      tags:
      - artist
    patch:
      consumes:
      - application/json
      description: Изменить данные исполнителя.
      parameters:
      - in: path
        name: artist-id
        required: true
        type: integer
      - description: Новые данные исполнителя
        in: body
        name: artist
        required: true
        schema:
          $ref: '#/definitions/artistrest.ChangeArtistRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/artistrest.ChangeArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Изменить данные исполнителя.
      tags:
      - artist
  /songs/:
    get:
      consumes:
      - application/json
      description: Поиск определенной песни по всем атрибутам.
      parameters:
      - in: query
        name: artistName
        type: string
      - in: query
        name: link
        type: string
      - in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/songrest.SearchSongsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Поиск определенной песни.
      tags:
      - song
    post:
      consumes:
      - application/json
      description: Добавление новой песни. Для разделения куплетов необходимо использовать
        '\n\n'.
      parameters:
      - description: Данные новой песни
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/songrest.CreateSongRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/songrest.CreateSongResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Добавить новую песню.
      tags:
      - song
  /songs/{song-id}:
    delete:
      consumes:
      - application/json
      description: Удалить данные песни.
      parameters:
      - in: path
        name: song-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/songrest.RemoveSongResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Удалить данные песни.
      tags:
      - song
    patch:
      consumes:
      - application/json
      description: Изменить данные песни. Для разделения куплетов необходимо использовать
        '\n\n'.
      parameters:
      - in: path
        name: song-id
        required: true
        type: integer
      - description: Новые данные песни
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/songrest.ChangeSongRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/songrest.ChangeSongResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Изменить данные песни.
      tags:
      - song
  /songs/{song-id}/couplets:
    get:
      consumes:
      - application/json
      description: Получить данные определенной песни с пагинацией по куплетами.
      parameters:
      - in: path
        name: song-id
        required: true
        type: integer
      - in: query
        minimum: 1
        name: pageNumber
        type: integer
      - in: query
        maximum: 100
        minimum: 10
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/songrest.GetSongResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mwerror.ErrorResponse'
      summary: Получить данные определенной песни.
      tags:
      - song
swagger: "2.0"